Rename wind_cube_nwtc to wind_cube_profile
