+++ ingest/wind_cube_profile/config/pipeline_config_nwtc_rtd.yml
@@ -3,7 +3,7 @@ pipeline:
 
   # These parameters will be used to name files.
   location_id: "nwtc"
-  dataset_name: "wind_cube_nwtc_sta"
+  dataset_name: "wind_cube_profile_rtd"
   # qualifier: ""
   # temporal: ""
   data_level: "b0" # If not applying QC this should be set to "a0"
@@ -26,7 +26,7 @@ dataset_definition:
   variables:
     time:
       input:
-        name: Timestamp (end of interval)
+        name: Timestamp
         converter:
           classname: tsdat.utils.converters.StringTimeConverter
           parameters:
@@ -84,14 +84,14 @@ dataset_definition:
         long_name: Wiper Count
         units: count
 
-    # Vbatt:
-    #   dims: [time]
-    #   type: float
-    #   input:
-    #     name: Vbatt (V)
-    #   attrs:
-    #     long_name: Vbatt
-    #     units: V
+    Vbatt:
+      dims: [time]
+      type: float
+      input:
+        name: Vbatt (V)
+      attrs:
+        long_name: Vbatt
+        units: V
 
     height:
       # data: [41, 55, 61, 74, 80, 87, 92, 100, 105, 119, 122, 130]
@@ -112,14 +112,14 @@ dataset_definition:
         long_name: Wind Speed
         units: m/s
 
-    # wind_speed_rad:
-    #   dims: [time, height]
-    #   type: float
-    #   input:
-    #     name: wind_speed_rad
-    #   attrs:
-    #     long_name: Radial Wind Speed
-    #     units: m/s
+    wind_speed_rad:
+      dims: [time, height]
+      type: float
+      input:
+        name: wind_speed_rad
+      attrs:
+        long_name: Radial Wind Speed
+        units: m/s
 
     wind_speed_disp:
       dims: [time, height]
@@ -130,23 +130,19 @@ dataset_definition:
         long_name: Radial Wind Speed Dispersion
         units: m/s
 
-    wind_speed_min:
-      dims: [time, height]
-      type: float
-      input:
-        name: wind_speed_min
-      attrs:
-        long_name: Wind Speed Min
-        units: m/s
+    # wind_speed_min:
+    #   dims: [time, height]
+    #   type: float
+    #   attrs:
+    #     long_name: Wind Speed Min
+    #     units: m/s
 
-    wind_speed_max:
-      dims: [time, height]
-      type: float
-      input:
-        name: wind_speed_max
-      attrs:
-        long_name: Wind Speed Max
-        units: m/s
+    # wind_speed_max:
+    #   dims: [time, height]
+    #   type: float
+    #   attrs:
+    #     long_name: Wind Speed Max
+    #     units: m/s
 
     wind_speed_z:
       dims: [time, height]
@@ -157,32 +153,23 @@ dataset_definition:
         long_name: z-wind
         units: m/s
 
-    wind_speed_z_disp:
+    wind_speed_x:
       dims: [time, height]
       type: float
       input:
-        name: wind_speed_z_disp
+        name: wind_speed_x
       attrs:
-        long_name: Z-wind Dispersion
+        long_name: x-wind
         units: m/s
 
-    # wind_speed_x:
-    #   dims: [time, height]
-    #   type: float
-    #   input:
-    #     name: wind_speed_x
-    #   attrs:
-    #     long_name: x-wind
-    #     units: m/s
-
-    # wind_speed_y:
-    #   dims: [time, height]
-    #   type: float
-    #   input:
-    #     name: wind_speed_y
-    #   attrs:
-    #     long_name: y-wind
-    #     units: m/s
+    wind_speed_y:
+      dims: [time, height]
+      type: float
+      input:
+        name: wind_speed_y
+      attrs:
+        long_name: y-wind
+        units: m/s
 
     # z_wind_dispersion:
     #   dims: [time, height]
@@ -209,40 +196,34 @@ dataset_definition:
         long_name: Carrier-to-noise ratio
         units: dB
 
-    cnr_min:
-      dims: [time, height]
-      type: float
-      input:
-        name: cnr_min
-      attrs:
-        long_name: CRN_Min
-        units: dB
-
-    dopp_spect_broad:
-      dims: [time, height]
-      type: float
-      input:
-        name: dopp_spect_broad
-      attrs:
-        long_name: Dopp Spect Broad
-        units: m/s
-
-    data_availability:
-      dims: [time, height]
-      type: float
-      input:
-        name: data_availability
-      attrs:
-        long_name: Availability
-        units: "%"
-
-    # position:
-    #   dims: [time]
-    #   type: int
-    #   input:
-    #     name: Position
+    # crn_min:
+    #   dims: [time, height]
+    #   type: float
     #   attrs:
-    #     description: Azimuthal position of scan. V is replaced with -1 for a vertical scan.
+    #     long_name: CRN_Min
+    #     units: dB
+
+    # dopp_spect_broad:
+    #   dims: [time, height]
+    #   type: float
+    #   attrs:
+    #     long_name: Dopp Spect Broad
+    #     units: m/s
+
+    # data_availability:
+    #   dims: [time, height]
+    #   type: float
+    #   attrs:
+    #     long_name: Availability
+    #     units: "%"
+
+    position:
+      dims: [time]
+      type: int
+      input:
+        name: Position
+      attrs:
+        description: Azimuthal position of scan. V is replaced with -1 for a vertical scan.
 
     latitude:
       data: 71.323
@@ -312,6 +293,6 @@ quality_management:
 
   # manage_custom_qc: # Rename this
   #   checker:
-  #     classname: ingest.wind_cube_nwtc.pipeline.qc.CustomQualityChecker
+  #     classname: ingest.wind_cube_profile.pipeline.qc.CustomQualityChecker
   #   handlers:
-  #     - classname: ingest.wind_cube_nwtc.pipeline.qc.CustomQualityHandler
+  #     - classname: ingest.wind_cube_profile.pipeline.qc.CustomQualityHandler
