+++ ingest/wind_cube_profile/pipeline/pipeline.py
