+++ ingest/wind_cube_profile/pipeline/styling.mplstyle
