+++ ingest/wind_cube_profile/tests/data/expected/nwtc.wind_cube_profile_rtd.b0.20220201.000000.nc
