+++ ingest/wind_cube_profile/tests/data/expected/nwtc.wind_cube_profile_sta.b0.20220202.000100.nc
