+++ ingest/wind_cube_profile/tests/data/input/WLS71460_2022_02_01__00_00_00.rtd.7z
