+++ ingest/wind_cube_profile/tests/data/input/WLS71460_2022_02_02__00_01_00.sta.7z
